Spin off consensus apis from block apis in json_rpc/block.js
