+++ json_rpc/block.js
@@ -66,37 +66,10 @@ module.exports = function getBlockApis(node) {
       done(null, JsonRpcUtil.addProtocolVersion({ result: blockHeaders }));
     },
 
-    [JSON_RPC_METHODS.AIN_GET_PROPOSER_BY_HASH]: function(args, done) {
-      const beginTime = Date.now();
-      const block = node.bc.getBlockByHash(args.hash);
-      const result = block ? block.proposer : null;
-      const latency = Date.now() - beginTime;
-      trafficStatsManager.addEvent(TrafficEventTypes.JSON_RPC_GET, latency);
-      done(null, JsonRpcUtil.addProtocolVersion({ result }));
-    },
-
-    [JSON_RPC_METHODS.AIN_GET_PROPOSER_BY_NUMBER]: function(args, done) {
+    [JSON_RPC_METHODS.AIN_GET_BLOCK_TRANSACTION_COUNT_BY_NUMBER]: function(args, done) {
       const beginTime = Date.now();
       const block = node.bc.getBlockByNumber(args.number);
-      const result = block ? block.proposer : null;
-      const latency = Date.now() - beginTime;
-      trafficStatsManager.addEvent(TrafficEventTypes.JSON_RPC_GET, latency);
-      done(null, JsonRpcUtil.addProtocolVersion({ result }));
-    },
-
-    [JSON_RPC_METHODS.AIN_GET_VALIDATORS_BY_NUMBER]: function(args, done) {
-      const beginTime = Date.now();
-      const block = node.bc.getBlockByNumber(args.number);
-      const result = block ? block.validators : null;
-      const latency = Date.now() - beginTime;
-      trafficStatsManager.addEvent(TrafficEventTypes.JSON_RPC_GET, latency);
-      done(null, JsonRpcUtil.addProtocolVersion({ result }));
-    },
-
-    [JSON_RPC_METHODS.AIN_GET_VALIDATORS_BY_HASH]: function(args, done) {
-      const beginTime = Date.now();
-      const block = node.bc.getBlockByHash(args.hash);
-      const result = block ? block.validators : null;
+      const result = block ? block.transactions.length : null;
       const latency = Date.now() - beginTime;
       trafficStatsManager.addEvent(TrafficEventTypes.JSON_RPC_GET, latency);
       done(null, JsonRpcUtil.addProtocolVersion({ result }));
@@ -105,15 +78,6 @@ module.exports = function getBlockApis(node) {
     [JSON_RPC_METHODS.AIN_GET_BLOCK_TRANSACTION_COUNT_BY_HASH]: function(args, done) {
       const beginTime = Date.now();
       const block = node.bc.getBlockByHash(args.hash);
-      const result = block ? block.transactions.length : null;
-      const latency = Date.now() - beginTime;
-      trafficStatsManager.addEvent(TrafficEventTypes.JSON_RPC_GET, latency);
-      done(null, JsonRpcUtil.addProtocolVersion({ result }));
-    },
-
-    [JSON_RPC_METHODS.AIN_GET_BLOCK_TRANSACTION_COUNT_BY_NUMBER]: function(args, done) {
-      const beginTime = Date.now();
-      const block = node.bc.getBlockByNumber(args.number);
       const result = block ? block.transactions.length : null;
       const latency = Date.now() - beginTime;
       trafficStatsManager.addEvent(TrafficEventTypes.JSON_RPC_GET, latency);
